Removed specified locations at top of original file
In the original file there was a location entry that specified an AWS
region and ami type.  This overwrote the regions specified in the
services.

When these locations were added then the BP would deploy only one
dynamic web cluster.  Removing it and relying on the locations
specified under DynamicRegionsFabric now causes the BP to behave
correctly e.g. 1 web cluster in AWS, 1 in Azure